moderator can be of one or more hastags

+++ routes/taglist.js
@@ -1,9 +1,10 @@
 const app = require('express').Router();
 const bodyParser = require('body-parser');
-const { tagList, hashtagFollowers, profiles } = require('../models');
+const { tagList, hashtagFollowers, profiles, hashtagModerators } = require('../models');
 const { Op } = require('sequelize');
-const { checkjwt, authorized, authorizedForProfileUUID, checkActiveUUID } = require('../middleware/jwtcheck');
+const { checkjwt, checkActiveUUID, authorized } = require('../middleware/jwtcheck');
 const { nullCheck, defaultNullFields } = require('./validations/nullcheck');
+const { authorizedAsModerator } = require('./../middleware/rolecheck');
 
 app.use(bodyParser.json());
 
@@ -11,12 +12,15 @@ app.use(bodyParser.json());
 * /:uuid - POST - create a tag
 * @check check jwt signature, match uuid of url with payload, check uuid from txt file
 */
-app.post("/:uuid", checkjwt, authorized, checkActiveUUID, async (req, res) => {
+app.post("/:uuid", checkjwt, authorizedAsModerator, checkActiveUUID, async (req, res) => {
     value = nullCheck(req.body, {
         nonNullableFields: ['tag', 'image', 'color', 'description'],
         mustBeNullFields: [...defaultNullFields, 'count', 'followerCount']
     });
     if (typeof (value) == 'string') return res.status(409).send(value);
+
+    const result = roleCheck(uuid, 'moderator');
+    if (typeof (result) == 'string') return res.status(403).send(result);
 
     await tagList.create(req.body)
         .then((result) => {
@@ -28,11 +32,38 @@ app.post("/:uuid", checkjwt, authorized, checkActiveUUID, async (req, res) => {
 });
 
 /* 
+* /:uuid - PUT - update a tag
+* @check check jwt signature, match uuid of url with payload, check uuid from txt file
+*/
+app.put("/:uuid", checkjwt, authorizedAsModerator, checkActiveUUID, async (req, res) => {
+    value = nullCheck(req.body, {
+        nonNullableFields: ['image', 'color', 'description'],
+        mustBeNullFields: [...defaultNullFields, 'count', 'followerCount', 'tag']
+    });
+    if (typeof (value) == 'string') return res.status(409).send(value);
+
+    const result = roleCheck(uuid, 'moderator');
+    if (typeof (result) == 'string') return res.status(403).send(result);
+
+    await tagList.update(req.body)
+        .then((result) => {
+            res.send("tag updated successfully!!");
+        })
+        .catch((err) => {
+            res.status(403).send(err);
+        });
+});
+
+/* 
 * /:tagUUID - DELETE - delete a tag
 * @check check jwt signature
 */
-app.delete("/:tagUUID", checkjwt, async (req, res) => {
+app.delete("/:tagUUID", checkjwt, authorizedAsModerator, async (req, res) => {
     const tagUUID = req.params.tagUUID;
+
+    const result = roleCheck(req.auth.uuid, 'admin');
+    if (typeof (result) == 'string') return res.status(403).send(result);
+
     await tagList.destroy({
         where: {
             "uuid": {
@@ -269,6 +300,120 @@ app.delete("/:profileUUID/unfollows/:tagUUID", checkjwt, authorizedForProfileUUI
     })
         .then((result) => {
             res.send("hashtag unfollowed successfully!!");
+        })
+        .catch((err) => {
+            res.status(403).send(err);
+        });
+});
+
+/* 
+* /:tagUUID/moderator/:uuid - POST - add moderator in hashtag
+* @check check jwt signature, match uuid from url with payload, check uuid in txt file
+*/
+app.post("/:tagUUID/moderator/:uuid", checkjwt, authorized, authorizedAsModerator, checkActiveUUID, async (req, res) => {
+    const tagUUID = req.params.tagUUID;
+    const uuid = req.params.uuid;
+    let error = false;
+
+    result = await tagList.findOne({
+        where: {
+            "uuid": {
+                [Op.eq]: tagUUID,
+            },
+        },
+        attributes: ['id'],
+    })
+        .catch((err) => {
+            error = true;
+            res.status(403).send(err);
+        });
+
+    if (error) return;
+
+    const hashtagId = result.id;
+
+    result = await users.findOne({
+        where: {
+            "uuid": {
+                [Op.eq]: uuid,
+            },
+        },
+        attributes: ['id'],
+    })
+        .catch((err) => {
+            error = true;
+            res.status(403).send(err);
+        });
+
+    if (error) return;
+
+    const userId = result.id;
+
+    await hashtagModerators.create({
+        "hashtagId": hashtagId,
+        "userId": userId,
+    })
+        .then((result) => {
+            res.send("moderator added successfully!!");
+        })
+        .catch((err) => {
+            res.status(403).send(err);
+        });
+});
+
+/* 
+* /:tagUUID/moderator/:uuid - DELETE - remove moderator in hashtag
+* @check check jwt signature, match uuid from url with payload, check uuid in txt file
+*/
+app.delete("/:tagUUID/moderator/:uuid", checkjwt, authorized, authorizedAsModerator, checkActiveUUID, async (req, res) => {
+    const tagUUID = req.params.tagUUID;
+    const uuid = req.params.uuid;
+    let error = false;
+
+    result = await tagList.findOne({
+        where: {
+            "uuid": {
+                [Op.eq]: tagUUID,
+            },
+        },
+        attributes: ['id'],
+    })
+        .catch((err) => {
+            error = true;
+            res.status(403).send(err);
+        });
+
+    if (error) return;
+
+    const hashtagId = result.id;
+
+    result = await users.findOne({
+        where: {
+            "uuid": {
+                [Op.eq]: uuid,
+            },
+        },
+        attributes: ['id'],
+    })
+        .catch((err) => {
+            error = true;
+            res.status(403).send(err);
+        });
+
+    if (error) return;
+
+    const userId = result.id;
+
+    await hashtagModerators.destroy({
+        "hashtagId": {
+            [Op.eq]: hashtagId
+        },
+        "userId": {
+            [Op.eq]: userId
+        },
+    })
+        .then((result) => {
+            res.send("moderator added successfully!!");
         })
         .catch((err) => {
             res.status(403).send(err);
